requirements/production.txt
django
djangorestframework
django-filter
drf-yasg
Pillow